refactor: centralize supabase client
