+++ frontend/src/hooks/useAuth.tsx
@@ -1,6 +1,6 @@
 import { createContext, useContext, useEffect, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
-import { supabase } from '../supabase'
+import { supabase } from '../lib/supabase'
 import { onAuthStateChange } from '../utils/auth'
 
 interface AuthContextProps {
