+++ frontend/src/pages/ConfigSettings.jsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import { supabase } from '../supabase'
+import { supabase } from '../lib/supabase'
 
 export default function ConfigSettings() {
   const [apiUrl, setApiUrl] = useState('')
