+++ frontend/src/pages/Home.jsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react'
-import { supabase } from '../supabase'
+import { supabase } from '../lib/supabase'
 import { useRole } from '../RoleContext'
 
 export default function Home({ onViewReports }) {
