+++ frontend/src/pages/Shifts.jsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react'
-import { supabase } from '../supabase'
+import { supabase } from '../lib/supabase'
 import { addShift } from '../supabase/shifts'
 
 export default function Shifts() {
