--- frontend/src/supabase.js
@@ -1,12 +1,0 @@
-import { createClient } from '@supabase/supabase-js'
-
-const supabaseUrl = import.meta.env.VITE_SUPABASE_URL
-const supabaseKey = import.meta.env.VITE_SUPABASE_ANON_KEY
-
-if (!supabaseUrl || !supabaseKey) {
-  console.error('Supabase environment variables are missing')
-}
-
-const supabase = createClient(supabaseUrl, supabaseKey)
-
-export { supabase }
