+++ frontend/src/supabase/dailyReports.js
@@ -1,4 +1,4 @@
-import { supabase } from '../supabase'
+import { supabase } from '../lib/supabase'
 
 export async function getReports() {
   const { data, error } = await supabase
