+++ frontend/src/supabase/inventory.js
@@ -1,4 +1,4 @@
-import { supabase } from '../supabase'
+import { supabase } from '../lib/supabase'
 
 const LOW_STOCK_THRESHOLD = 5
