+++ frontend/src/supabase/orders.js
@@ -1,4 +1,4 @@
-import { supabase } from '../supabase'
+import { supabase } from '../lib/supabase'
 
 export async function getOrders() {
   const { data, error } = await supabase
