+++ frontend/src/supabase/shifts.js
@@ -1,4 +1,4 @@
-import { supabase } from '../supabase'
+import { supabase } from '../lib/supabase'
 
 export async function addShift(data) {
   const { error } = await supabase.from('shifts').insert(data)
