+++ frontend/src/supabase/shiftsSchedule.js
@@ -1,4 +1,4 @@
-import { supabase } from '../supabase'
+import { supabase } from '../lib/supabase'
 
 export async function getSchedule() {
   const { data, error } = await supabase
