+++ frontend/src/supabase/upsell.js
@@ -1,4 +1,4 @@
-import { supabase } from '../supabase'
+import { supabase } from '../lib/supabase'
 
 export async function getUpsellItems() {
   const { data, error } = await supabase
