+++ frontend/src/utils/auth.ts
@@ -1,8 +1,5 @@
-import { createClient, Session, User } from '@supabase/supabase-js'
-
-const supabaseUrl = import.meta.env.VITE_SUPABASE_URL
-const supabaseKey = import.meta.env.VITE_SUPABASE_ANON_KEY
-export const supabase = createClient(supabaseUrl, supabaseKey)
+import { Session, User } from '@supabase/supabase-js'
+import { supabase } from '../lib/supabase'
 
 export async function signIn(email: string, password: string) {
   const { data, error } = await supabase.auth.signInWithPassword({ email, password })
